+++ src/c/HeartBeatView.h
@@ -5,7 +5,7 @@
 #include "ViewSelector.h"
 
 void initLayoutHeartBeat();
-void updateHeartBeat(int32_t value);
+void updateHeartBeat(int8_t value);
 void updateHeartBeatHistory();
 void drawHeartBeat(Layer *frame, GContext* context);
